refactor + integrate sync  with mesh.Block
