+++ sync/sync_test.go
@@ -4,11 +4,46 @@ import (
 	"fmt"
 	"github.com/gogo/protobuf/proto"
 	"github.com/spacemeshos/go-spacemesh/mesh"
+	"github.com/spacemeshos/go-spacemesh/p2p"
 	"github.com/spacemeshos/go-spacemesh/sync/pb"
 	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
 )
+
+type PeersMocks struct {
+	p2p.Service
+}
+
+func (ml PeersMocks) Count() int {
+	return 10
+}
+
+func (ml PeersMocks) LatestLayer() int {
+	return 10
+}
+
+func (ml PeersMocks) GetLayerHash(peer int) string {
+	return "asiodfu45987345" //some random string
+}
+
+func (ml PeersMocks) GetBlockByID(peer Peer, id string) (Block, error) {
+	return nil, nil
+}
+
+func (ml PeersMocks) ChoosePeers(pNum int) []Peer {
+	return nil
+}
+
+func (ml PeersMocks) GetLayerBlockIDs(peers Peer, i int, hash string) ([]string, error) {
+
+	return nil, nil
+}
+
+func (ml PeersMocks) GetPeers() []Peer {
+
+	return nil
+}
 
 func NewBlockResponseHandler() (func(msg []byte), chan *pb.FetchBlockResp) {
 	ch := make(chan *pb.FetchBlockResp)
